feat: add pagination to software and services admin dashboard tables (#128)
this adds pagination to software and services admin dashboard tables.
this should also improve the table lagginess a bit, which was pretty bad
for the services table with ~350 rows. reconsider this once react-aria's
table has virtualisation support. current page size is 50 rows,
pagination state is currently not synced to url search params.