fix: add workaround for datefield validation (#286)
the date field from `react-aria-components` currently validates each
date segment immediately when numbers are entered. internally, it always
requires a valid date, and defaults to the current date. therefore, when
we are currently in a month with 30 days, and a user starts entering a
date in the date field which starts with 31, it will not allow that.

the current workaround, until this is fixed upstream, is to explicitly
set a placeholder date, which defaults to 2024-01-01 (january has 31
days, and 2024 was a leap year).

closes #283