feat: derive working group permissions from contributions (#328)
- adds an optional relation from user to person.
- derives user permissions as they relate to working groups from entries
in the contributions table.
- adds a /dashboard/working-groups/:id route for testing

note: many contributions (e.g. for wg_chair) do not have a start date.
we currently only grant permissions when a start date is set and is in
the past. this may require relaxing in the future depending on how hard
it is to gather that information.