ci: use anubis to filter out bot traffic (#349)
this disables the "ingress" setting via `auto-deploy-values.yaml` (used
by our deployment workflow), so we don't overwrite the manual setup in
rancher, which has configured `anubis` at the ingress level to filter
out unwanted bot traffic.

when we change the deployment domain in the future, we need to enable
the "ingress" setting again (see the template
[here](https://github.com/acdh-oeaw/gl-autodevops-minimal-port/tree/main?tab=readme-ov-file#customizing-the-deployment)),
re-deploy, reconfigure `anubis` in `rancher`, and then disable the
"ingress" setting in `auto-deploy-values.yaml` again.